CONSOLIDATE ALL THREE PROMPT SAMPLERS INTO MOLMOSPACES.UTILS.PROMPT_SAMPLER
Three overlapping "how do we phrase a task instruction" implementations
existed, scattered: g1_molmo_port's own gold-exact PromptSampler
(components/prompt_sampler.py), an unqualified PromptSampler in
policy/learned_policy/utils.py (the real production one, feeding learned/
VLM policies), and ObjectMeta.get_short_description/clean_object_name
(utils/object_metadata.py, whose own TODOs already say "ported from
PromptSampler late at night").

Moved both actual sampler classes to sit next to each other in
molmo_spaces/utils/prompt_sampler.py, renamed to describe what
distinguishes them instead of colliding on the same name:
- PromptSamplerSimple: gold's own template port (rng-random selection,
  gzip-JSON objathor metadata lookup). Call sites (g1_molmo_port/tasks/
  pick_g1ms.py, open.py) updated, behavior unchanged.
- PromptSamplerLearnedPolicy: the real production sampler (renamed from
  the unqualified PromptSampler in policy/learned_policy/utils.py),
  moved verbatim -- cyclic .next() rotation (not rng-random),
  ObjectMeta-backed, custom-object-name override, distractor
  disambiguation by position, pick_and_place two-slot template. Confirmed
  uncalled anywhere in the codebase both before and after this move.

policy/learned_policy/utils.py lost the class plus its now-unused
imports (ObjectMeta, BaseMujocoTask, Rotation, logging); left a pointer
comment to the new location. Fixed a stale comment reference in
scripts/benchmarks/patch_task_descriptions.py.

Not addressed here: the still-unresolved overlap between these two and
ObjectMeta's own short-description lookup (gzip-JSON vs lmdb backend) --
noted in PromptSamplerSimple's docstring for a future pass.

Verified via generate_ported_rollout.py (seed 0, bit-exact vs gold) after
every step of this move.
