+++ scripts/benchmarks/patch_task_descriptions.py
@@ -5,7 +5,8 @@ For each task, resolves the pickup object's asset_id from the object name hash,
 then uses ObjectMeta.short_descriptions (1/2/3 word name) to set the
 task_description using the appropriate template based on task_cls.
 
-This matches the runtime behavior of PromptSampler.
+This matches the runtime behavior of PromptSamplerLearnedPolicy
+(molmo_spaces.utils.prompt_sampler).
 """
 
 import argparse
